Add server IP and port parameters to Api class methods
The API class methods 'instance', 'Api', and 'init' have been updated to include parameters for 'serverIp' and 'port'. This change was made not only to the function definitions located in 'Api.h', but also in the corresponding function implementations within 'Api.cpp'. This addition will allow for explicit server and port settings when initializing API instances.